requirements.txt
streamlit==1.36.0
python-dotenv==1.0.1

pandas==2.2.2
numpy==1.26.4

requests==2.32.3
tenacity==8.5.0

faiss-cpu==1.8.0.post1
sentence-transformers==3.0.1

psycopg2-binary==2.9.9

# Optionnel mais utile
rapidfuzz==3.9.6
tqdm==4.66.4
